Refactor styles
The styles were refactored so that the styling of the  customized
buttons are compatible with Firefox.

Changes in sync.js were made because of new class names for the
html elements.

Styles were added to the table data containing the syn direction
because the contents collapse when using templates other than
bootstrap.

Styles for color and fill were reapplied after accidental
removal.